move rotate tool into its own file
