+++ renderlib/Manipulator.h
@@ -154,7 +154,10 @@ struct Origins
     glm::vec3 dir = p - scene.scene->m_boundingBox.GetCenter();
 
     // now rotate dir by rotation
-    // TODO
+    // turn motion euler angles into quaternion
+    glm::quat q = glm::quat(rotation);
+    // transform dir by quaternion
+    dir = q * dir;
 
     // now set distance, theta, phi from m_P and current m_Target
     l.m_Distance = glm::length(dir);
@@ -212,35 +215,4 @@ struct AreaLightTool : ManipulationTool
 
   virtual void action(SceneView& scene, Gesture& gesture) final;
   virtual void draw(SceneView& scene, Gesture& gesture) final;
-};
-
-struct RotateTool : ManipulationTool
-{
-  // Selection codes, are used to identify which manipulator is under the cursor.
-  // The values in this enum are important, lower values means higher picking priority.
-  enum RotateCodes
-  {
-    kRotateX = 0,    // constrained to rotate about x axis
-    kRotateY = 1,    // constrained to rotate about y axis
-    kRotateZ = 2,    // constrained to rotate about z axis
-    kRotateView = 3, // constrained to rotate about view direction
-    kRotate = 4,     // general tumble rotation
-    kLast = 5
-  };
-
-  RotateTool()
-    : ManipulationTool(kLast)
-  {
-  }
-
-  virtual void action(SceneView& scene, Gesture& gesture) final;
-  virtual void draw(SceneView& scene, Gesture& gesture) final;
-
-  // Some data structure to store the initial state of the objects
-  // to move.
-  Origins origins;
-
-  // The current rotation of the objects to move.
-  // We need to potentially access this across calls to action and draw
-  glm::vec3 m_rotation;
 };
